move client-only generatePreviewHTML to its own module
