+++ client/lib/generateCommentHTML.ts
@@ -1,10 +1,4 @@
 import { JSDOM } from "jsdom";
-import rehypeStringify from "rehype-stringify";
-import remarkParse from "remark-parse";
-import remarkRehype from "remark-rehype";
-import { unified } from "unified";
-
-import { PreviewComment } from "~/types/client/utils.type";
 
 // Probably will integrate this to the build chain and compress it.
 const extraScript = (apiURL: string) => `const apiURL = "${apiURL}";
@@ -85,7 +79,7 @@ window.addEventListener("load", () => {
   validate();
 });`;
 
-export function generateCommentHTML(html: string, apiURL: string, isDark?: boolean) {
+export default function generateCommentHTML(html: string, apiURL: string, isDark?: boolean) {
     if (typeof window !== "undefined") throw new Error("This function should be ran on the server");
     const dom = new JSDOM(html);
     const document = dom.window.document;
@@ -94,47 +88,4 @@ export function generateCommentHTML(html: string, apiURL: string, isDark?: boole
     document.body.appendChild(script);
     if (isDark) document.documentElement.classList.add("dark");
     return dom.serialize();
-}
-
-export async function generatePreviewHTML(
-    html: string,
-    comments: PreviewComment[],
-    isDark?: boolean
-) {
-    if (typeof window === "undefined") throw new Error("This should be run on client side");
-    const remark = unified().use(remarkParse).use(remarkRehype).use(rehypeStringify);
-    const commentHTML = await Promise.all(
-        comments.map(async ({ author, content, date }) => ({
-            author,
-            date,
-            content: String(await remark.process(content)),
-        }))
-    );
-
-    const htmlDom = new DOMParser().parseFromString(html, "text/html");
-    const doc = htmlDom.documentElement;
-    const commentWrapper = doc.querySelector("[data-ezk=comments]");
-    if (!commentWrapper) return html;
-
-    const commentFrameHTML = commentWrapper.innerHTML;
-    commentWrapper.innerHTML =
-        commentHTML.length > 0
-            ? ""
-            : "<div>There are no comments yet. Be the first to join the conversation.</div>";
-
-    for (let i = 0; i < commentHTML.length; i++) {
-        const commentDom = new DOMParser().parseFromString(commentFrameHTML, "text/html");
-        const commentDoc = commentDom.documentElement;
-        const author = commentDoc.querySelector("[data-ezk='comment-author']");
-        const date = commentDoc.querySelector("[data-ezk='comment-date']");
-        const content = commentDoc.querySelector("[data-ezk='comment-content']");
-        if (author) author.textContent = commentHTML[i].author ?? "Anonymous";
-        if (date) date.textContent = commentHTML[i].date;
-        if (content) content.innerHTML = commentHTML[i].content;
-        commentWrapper.appendChild(commentDom.body.firstChild as ChildNode);
-    }
-
-    if (isDark) doc.classList.add("dark");
-
-    return new XMLSerializer().serializeToString(doc);
 }
